feat(commands): add filter option to main script
