+++ package.json
@@ -26,6 +26,7 @@
     "request": "^2.88.2",
     "spotify-web-api-node": "^5.0.2",
     "typescript": "^4.5.4",
+    "yargs": "^17.3.1",
     "zod": "^3.11.6"
   },
   "devDependencies": {
@@ -35,15 +36,16 @@
     "@types/node": "^17.0.9",
     "@types/request": "^2.48.8",
     "@types/spotify-web-api-node": "^5.0.6",
+    "@types/yargs": "^17.0.8",
     "ava": "^4.0.1",
     "ts-node": "^10.4.0"
   },
   "ava": {
-		"extensions": [
-			"ts"
-		],
-		"require": [
-			"ts-node/register"
-		]
-	}
+    "extensions": [
+      "ts"
+    ],
+    "require": [
+      "ts-node/register"
+    ]
+  }
 }
